Enable dynamic rendering
