+++ app/blog/[slug]/page.tsx
@@ -1,18 +1,8 @@
-import { getAllBlogPosts, getBlogPostBySlug } from "@/lib/content"
+export const dynamic = "force-dynamic"
+
+import { getBlogPostBySlug } from "@/lib/content"
 import { BlogPostPageClient } from "./BlogPostPageClient" // Existing client component
 import { notFound } from "next/navigation"
-
-export async function generateStaticParams() {
-  try {
-    const posts = await getAllBlogPosts()
-    return posts.map((post) => ({
-      slug: post.slug,
-    }))
-  } catch (error) {
-    console.error("Error generating static params for blog posts:", error)
-    return []
-  }
-}
 
 interface BlogPostPageProps {
   params: {
