+++ app/blog/page.tsx
@@ -1,4 +1,4 @@
-"use client"
+export const dynamic = "force-dynamic"
 
 import { Suspense } from "react"
 import { type BlogPost } from "@/lib/content"
